requirements.txt
anyio==3.7.0
appwrite==2.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
idna==3.4
pydantic==1.10.8
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.6.3
urllib3==2.0.2
uvicorn==0.22.0
